// src/app/core/models/iglesia.ts
export class Iglesia {
  id: number;
  nombre: string;
  idPastor: number;
  ubicacion: string;
  direccion: string;
  fundacion: string;
  superdni : string;
  presimni : string;
  presijni : string;
  diacentral : string;
  horacentralini : string;
  horacentralfin : string;
  diajni : string;
  horajniini  : string;
  horajnifin  : string;
  zona: string;
  imagePath?: string;
  pastorname?: string;
  imagePathPas?: string;
}